Area3D Door Script
Used Area3D Door Script and set player input to spacebar for now to engage with doors. Player input logic will be replaced instead with whether player has a key or not (aka correct flame color). Doors will open if the player is within the Area3D of the door and they press space bar. However, for some reason this does not work for doors position left to right, it only works for doors position top to bottom on map. Working on a fix.